implemented presentation to ui mapper business logic and updated repository score data type

+++ domain/src/main/java/com/na/sadapay/domain/githubrepositories/model/GithubRepositoryDomainModel.kt
@@ -5,6 +5,6 @@ data class GithubRepositoryDomainModel(
     val description: String,
     val url: String,
     val language: String,
-    val score: Float,
+    val score: Long,
     val author: GithubRepositoryAuthorDomainModel
 )
