+++ presentation/src/main/java/com/na/sadapay/presentation/githubrepositories/model/GithubRepositoryPresentationModel.kt
@@ -5,6 +5,6 @@ data class GithubRepositoryPresentationModel(
     val description: String,
     val url: String,
     val language: String,
-    val score: Float,
+    val score: Long,
     val author: GithubRepositoryAuthorPresentationModel
 )
